Move `as_completed_stop_and_raise_on_error` to `file_transaction.py`
